Docs

+++ app/src/main/java/ru/eco/automan/Converters.kt
@@ -13,8 +13,8 @@ class Converters {
     }
 
     @TypeConverter
-    fun dateToTimestamp(date: Date?): Int? {
-        return date?.time?.toInt()
+    fun dateToTimestamp(date: Date?): Long? {
+        return date?.time?.toLong()
     }
 
 }
